+++ packages/sui-domain/src/Entity.js
@@ -1,7 +1,7 @@
-import inherits from './inherits'
-import AnemicModel from './AnemicModel'
+import inherits from './inherits.js'
+import AnemicModel from './AnemicModel.js'
 
-const Entity = (function(_AnemicModel) {
+const Entity = (function (_AnemicModel) {
   inherits(Entity, _AnemicModel)
 
   function Entity() {