+++ packages/sui-hoc/src/avoidSSRRender.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react'
+import {Component} from 'react'
 
 export default Target =>
   class extends Component {