+++ packages/sui-js-compiler/test/server/index.js
@@ -8,13 +8,41 @@ const libPath = path.join(__dirname, 'lib')
 const libFilePath = path.join(libPath, 'example.js')
 
 describe('@s-ui/js-compiler', () => {
-  after(() => fs.remove(libPath))
-  before(() => fs.remove(libPath))
+  afterEach(() => fs.remove(libPath))
+  beforeEach(() => fs.remove(libPath))
 
-  it('compile a /src folder with a JavaScript with JSX file and output to /lib', async () => {
-    const {stdout} = await exec('npx sui-js-compiler', {
+  it('compiles a /src folder with a JavaScript with JSX file and output to /lib', async () => {
+    const {stdout} = await exec('node ../../index.js', {
       cwd: __dirname
     })
+
+    const compiledFilenames = await fs.readdir(libPath)
+
+    expect(compiledFilenames).to.eql(['example.js', 'example.test.js'])
+
+    expect(stdout).to.contain('[sui-js-compiler]')
+
+    const compiledFile = await fs.readFile(libFilePath, 'utf-8')
+
+    console.log(compiledFile)
+    expect(compiledFile).to.contain('react/jsx-runtime')
+    expect(compiledFile).to.contain('_jsx')
+
+    expect(compiledFile).to.contain('applyDecoratedDescriptor')
+    expect(compiledFile).to.contain('regeneratorRuntime')
+  })
+
+  it('when the "ignore" option exists, it exclude all the file matching the passed patterns', async () => {
+    const {stdout} = await exec(
+      'node ../../index.js --ignore="./src/**.test.js"',
+      {
+        cwd: __dirname
+      }
+    )
+
+    const compiledFilenames = await fs.readdir(libPath)
+
+    expect(compiledFilenames).to.eql(['example.js'])
 
     expect(stdout).to.contain('[sui-js-compiler]')
 
@@ -24,7 +52,6 @@ describe('@s-ui/js-compiler', () => {
     expect(compiledFile).to.contain('_jsx')
 
     expect(compiledFile).to.contain('applyDecoratedDescriptor')
-    expect(compiledFile).to.contain('createClass')
     expect(compiledFile).to.contain('regeneratorRuntime')
   })
 })