+++ packages/sui-pde/src/components/feature.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import useFeature from '../hooks/useFeature'
+import useFeature from '../hooks/useFeature.js'
 
 export default function Feature({
   children,