+++ packages/sui-polyfills/src/element-closest.js
@@ -11,7 +11,7 @@ if (typeof window !== 'undefined') {
   }
 
   if (!Element.prototype.closest) {
-    Element.prototype.closest = function(s) {
+    Element.prototype.closest = function (s) {
       var el = this
 
       do {