+++ packages/sui-test/CHANGELOG.md
@@ -1,5 +1,70 @@
 # CHANGELOG
 
+# 8.9.0 (2022-04-07)
+
+
+### Features
+
+* **packages/sui-test:** update commander dependency ([1667366](https://github.com/SUI-Components/sui/commit/166736675be1158b2d2857ce28a432fafe205bdb))
+
+
+
+# 8.7.0 (2022-03-30)
+
+
+
+# 8.6.0 (2022-03-28)
+
+
+### Features
+
+* **packages/sui-test:** Add new headless flag ([1f6fa24](https://github.com/SUI-Components/sui/commit/1f6fa24a11a6c08e47c30f1382178eb3b5a644bd))
+* **packages/sui-test:** Upgrade dependencies ([110e5eb](https://github.com/SUI-Components/sui/commit/110e5eb7f154454102701661932e55a05800532f))
+
+
+
+# 8.5.0 (2022-03-01)
+
+
+### Bug Fixes
+
+* **packages/sui-test:** Force alias for sui-test ([91bd068](https://github.com/SUI-Components/sui/commit/91bd0681569ae4aacbfa33011dae9f59b0416c57))
+
+
+
+# 8.4.0 (2022-02-22)
+
+
+### Features
+
+* **packages/sui-test:** Upgrade dependencies ([ba72bf4](https://github.com/SUI-Components/sui/commit/ba72bf47ad8c7ced3e195ca5b08bfe743edee07e))
+
+
+
+# 8.3.0 (2022-02-17)
+
+
+### Bug Fixes
+
+* **packages/sui-test:** fix extension ([0bf5334](https://github.com/SUI-Components/sui/commit/0bf53346ce549be38681a153274b2aef01a8ad2f))
+* **packages/sui-test:** fix some lint warnings ([1b16de8](https://github.com/SUI-Components/sui/commit/1b16de8cbc3653fd70c0f61196d5cdf7bea4f8f3))
+
+
+### Features
+
+* **packages/sui-test:** add describe on local ([64ef8ab](https://github.com/SUI-Components/sui/commit/64ef8ab91c71fe8fabba8e714b5330338d4c8d3d))
+
+
+
+# 8.2.0 (2022-02-16)
+
+
+### Features
+
+* **packages/sui-test:** Upgrade dependencies ([95e55d5](https://github.com/SUI-Components/sui/commit/95e55d56417dae5bd26300d9fdb911bf3896e554))
+
+
+
 # 8.1.0 (2022-01-31)